# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Telegram Drive API",
    "description": "HTTP API for Telegram-backed object storage. Headless (Docker) exposes the full surface on one port (1334): legacy /upload, /metrics, WebDAV, static Web UI, and /d/* share downloads. Desktop optional REST API (127.0.0.1:8550) is a subset — see docs/DESKTOP-API.md. Authenticate with X-API-Key or X-Access-Pwd (Web admin / desktop local access password).",
    "version": "3.0.0-beta"
  },
  "servers": [{ "url": "/", "description": "Current host" }],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "AccessPwdAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Access-Pwd",
        "description": "Web admin console password (ACCESS_PWD)"
      }
    },
...
  },
  "paths": {
...
    "/stream/{folder_id}/{message_id}": {
      "get": {
        "tags": ["Media"],
        "summary": "Stream media (token query param)",
        "parameters": [
          { "name": "folder_id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "message_id", "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "token", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Media stream" } }
      }
    }
  }
}